refactor(domain): 검증 로직이 필요한 name을 객체로 분리
길이 검증이 필요한 name을 객체로 분리했습니다.
